requirements.txt
kagglehub
python-dotenv
pandas
boto3
